refactor: deslop code
